sale badge,colour
hi
